requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-yasg==1.21.9
djangorestframework_simplejwt==5.4.0
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-filter==25.1